+++ DependencyInjection/c975LEmailExtension.php
@@ -1,0 +1,29 @@
+<?php
+/*
+ * (c) 2017: 975l <contact@975l.com>
+ * (c) 2017: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace c975L\EmailBundle\DependencyInjection;
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\Loader;
+
+class c975LEmailExtension extends Extension
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function load(array $configs, ContainerBuilder $container)
+    {
+        $configuration = new Configuration();
+        $processedConfig = $this->processConfiguration( $configuration, $configs );
+
+        $container->setParameter('c975_l_email.sentFrom', $processedConfig['sentFrom']);
+    }
+}