+++ Resources/sql/TableCreation.sql
@@ -15,19 +15,3 @@ CREATE TABLE `emails` (
   `ip` varchar(48) DEFAULT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb4;
-
--- ----------------------------
--- Table structure for emails_archives
--- ----------------------------
--- DROP TABLE IF EXISTS `emails_archives`;
-CREATE TABLE `emails_archives` (
-  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
-  `date_sent` datetime DEFAULT NULL,
-  `subject` varchar(256) DEFAULT NULL,
-  `sent_from` varchar(128) DEFAULT NULL,
-  `sent_to` varchar(128) DEFAULT NULL,
-  `sent_cc` varchar(128) DEFAULT NULL,
-  `body` text,
-  `ip` varchar(48) DEFAULT NULL,
-  PRIMARY KEY (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;