+++ ChangeLog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+v1.1
+----
+- Add of content in README.md (05/06/2017)
+- Add of configuration parameter
+
+v1.0.1
+------
+- Add of files (05/06/2017)
+
 v1.0
 ----
 - Creation of bundle  (05/06/2017)